backend/requirements.txt
Flask==3.0.0
langchain==0.0.353
llama_index==0.9.24
openai==1.6.1
pinecone_client==2.2.4
python-dotenv==1.0.0
